moved data types into a separate file, because the extra code to support pipelines made the main file too big, also testing to improve concurrency to get the lock to update the state of an event from a pipeline poin tof view
